Added some logs + admin

+++ src/main/java/com/storeManagementTool/StoreManagementTool/controllers/ProductController.java
@@ -2,6 +2,7 @@ package com.storeManagementTool.StoreManagementTool.controllers;
 
 import com.storeManagementTool.StoreManagementTool.dtos.ProductAddDTO;
 import com.storeManagementTool.StoreManagementTool.dtos.ProductDTO;
+import com.storeManagementTool.StoreManagementTool.entities.Role;
 import com.storeManagementTool.StoreManagementTool.entities.UserEntity;
 import com.storeManagementTool.StoreManagementTool.services.ProductService;
 import jakarta.annotation.Nullable;
@@ -22,32 +23,54 @@ public class ProductController {
     }
 
     @PostMapping
-    public ResponseEntity<ProductDTO> createProduct(@RequestBody ProductAddDTO productAddDTO) { /// TODO: case with negatives values for q/
+    public ResponseEntity<ProductDTO> createProduct(@RequestBody ProductAddDTO productAddDTO,
+                                                    @AuthenticationPrincipal UserEntity user) { /// TODO: case with negatives values for q/
+        if (user.getRole() != Role.ADMIN) {
+            log.error("Only admins can create products");
+            throw new IllegalArgumentException("Only admins can create products");
+        }
+
+        if (productAddDTO.getPrice() < 0 || productAddDTO.getQuantity() < 0) {
+            log.error("Price or quantity are negative");
+            throw new IllegalArgumentException("Price or quantity must be greater than 0");
+        }
+
         return ResponseEntity.ok(productService.save(productAddDTO));
     }
 
     @PutMapping
-    public ResponseEntity<ProductDTO> updateProduct(@RequestBody ProductDTO productDTO) {
+    public ResponseEntity<ProductDTO> updateProduct(@RequestBody ProductDTO productDTO,
+                                                    @AuthenticationPrincipal UserEntity user) {
+        if (user.getRole() != Role.ADMIN) {
+            log.error("Only admins can update products");
+            throw new IllegalArgumentException("Only admins can create products");
+        }
         return ResponseEntity.ok(productService.update(productDTO));
     }
 
     @PatchMapping({"/changePrice/{id}"})
     public ResponseEntity<ProductDTO> changePrice(@RequestParam Double price, @PathVariable Long id) {
         if( price < 0) {
+            log.error("Price is negative");
             throw new IllegalArgumentException("Price must be greater than 0");
         }
         return ResponseEntity.ok(productService.changePriceById(id, price));
     }
 
     @PatchMapping({"/changeQuantity/{id}"})
-    public ResponseEntity<ProductDTO> changeQuantity(@RequestParam Integer quantity, @PathVariable Long id) {
+    public ResponseEntity<ProductDTO> changeQuantity(@RequestParam Integer quantity, @PathVariable Long id,
+                                                     @AuthenticationPrincipal UserEntity user) {
+        if (user.getRole() != Role.ADMIN) {
+            log.error("Only admins can change quantity");
+            throw new IllegalArgumentException("Only admins can change quantity");
+        }
+
         if( quantity < 0) {
+            log.error("Quantity is negative");
             throw new IllegalArgumentException("Quantity must be greater than 0");
         }
         return ResponseEntity.ok(productService.changeQuantityById(id, quantity));
     }
-
-    //TODO: To test what happens when I try to update a product that don't exists
 
     @GetMapping
     public ResponseEntity<List<ProductDTO>> getAllProducts(@RequestParam @Nullable String name,
@@ -68,7 +91,10 @@ public class ProductController {
     }
 
     @DeleteMapping({"/{id}"})
-    public ResponseEntity<String> deleteById(@PathVariable Long id) {
+    public ResponseEntity<String> deleteById(@PathVariable Long id, @AuthenticationPrincipal UserEntity user) {
+        if (user.getRole() != Role.ADMIN) {
+            throw new IllegalArgumentException("Only admins can create products");
+        }
         productService.delete(id);
         return ResponseEntity.ok("Product with id " + id + " deleted successfully.");
     }
